spilt into rewrite_dependent_join & decorrelate_dependent_join
